# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Pizza Shop API",
        "description": "API Documentation for Pizza Shop",
        "contact": {
            "email": "support@pizzashop.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Main API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Register/Login"
                ],
                "summary": "Аутентификация пользователя",
                "operationId": "09c1cdfe33bfbe0710861ac50c0adaec",
                "requestBody": {
                    "description": "Данные для входа",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Неверные учетные данные"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email обязателен для заполнения"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Email обязателен для заполнения"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Register/Login"
                ],
                "summary": "Создание нового пользователя",
                "operationId": "08895790736612da345bc0606cefe61b",
                "requestBody": {
                    "description": "Данные пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "ФИО пользователя",
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "ivan@example.com"
                                    },
                                    "password": {
                                        "description": "Пароль (минимум 8 символов, только английские буквы и цифры)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "description": "Подтверждение пароля",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdef123456"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error: Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email обязателен для заполнения"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Email обязателен для заполнения"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Список продуктов",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Получить продукт по ID",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ресурс не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Запрашиваемый ресурс не существует"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получить данные аутентифицированного пользователя",
                "description": "Возвращает информацию о текущем пользователе, аутентифицированном через Sanctum",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Cart": {
                "required": [
                    "user_id",
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID пользователя, которому принадлежит корзина",
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "description": "ID продукта в корзине",
                        "type": "integer",
                        "example": 5
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "quantity": {
                        "description": "Количество товара",
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "description": "Дата создания записи",
                        "type": "string",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата последнего обновления",
                        "type": "string",
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Category": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название категории",
                        "type": "string",
                        "example": "Пицца"
                    }
                },
                "type": "object"
            },
            "Order": {
                "required": [
                    "user_id",
                    "phone",
                    "email",
                    "delivery_address"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID пользователя",
                        "type": "integer",
                        "format": "int64",
                        "example": 4
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "phone": {
                        "description": "Номер телефона",
                        "type": "string",
                        "example": "+79214074603"
                    },
                    "email": {
                        "description": "Электронная почта",
                        "type": "string",
                        "format": "email",
                        "example": "denis.fad@example.com"
                    },
                    "delivery_address": {
                        "description": "Адрес доставки",
                        "type": "string",
                        "example": "Улица Шевченко дом 3 корпус 2, кв 92 этаж 9"
                    },
                    "status_id": {
                        "description": "ID статуса заказа",
                        "type": "integer",
                        "format": "int64",
                        "example": 4
                    },
                    "status": {
                        "$ref": "#/components/schemas/Status"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "required": [
                    "order_id",
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "order_id": {
                        "description": "ID заказа",
                        "type": "integer",
                        "format": "int64",
                        "example": 24
                    },
                    "order": {
                        "$ref": "#/components/schemas/Order"
                    },
                    "product_id": {
                        "description": "ID продукта",
                        "type": "integer",
                        "format": "int64",
                        "example": 5
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "quantity": {
                        "description": "Колличество данного продукта",
                        "type": "integer",
                        "format": "int64",
                        "example": 9
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "ID категории продукта",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "name": {
                        "description": "Название продукта",
                        "type": "string",
                        "example": "Пепперони"
                    },
                    "description": {
                        "description": "Описание продукта",
                        "type": "string",
                        "example": "Пицца с колбасками пепперони и сыром"
                    },
                    "price": {
                        "description": "Цена продукта",
                        "type": "number",
                        "format": "float",
                        "example": 12.99
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "deleted_at": {
                        "description": "Дата удаления (мягкое удаление)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Status": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Название статуса",
                        "type": "string",
                        "example": "В обработке"
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Имя/Фамилия",
                        "type": "string",
                        "example": "Денис Фадыров"
                    },
                    "email": {
                        "description": "Электронная почта",
                        "type": "string",
                        "format": "email",
                        "example": "denis.fad@example.com"
                    },
                    "password": {
                        "description": "Пароль",
                        "type": "string",
                        "format": "password",
                        "example": "secret123"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Используйте Bearer Token для аутентификации. Получите токен через /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Register/Login",
            "description": "API Endpoint for Register/Login"
        },
        {
            "name": "Products",
            "description": "API Endpoints for Products"
        },
        {
            "name": "User",
            "description": "Операции с текущим аутентифицированным пользователем"
        }
    ]
}